Including the Gate in the methods

+++ app/Http/Controllers/RoleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Permissions\Models\Permission;
 use App\Permissions\Models\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -15,6 +16,7 @@ class RoleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('have_access', 'role.index');
         $roles = Role::orderBy('id', 'Desc')->paginate(2);
         return view('role.index', compact('roles'));
     }
@@ -26,6 +28,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('have_access', 'role.create');
         $permissions = Permission::get();
         return view('role.create', compact('permissions'));
     }
@@ -38,6 +41,7 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('have_access', 'role.create');
         $request->validate([
             'name' => 'required|max:50|unique:roles,name',
             'slug' => 'required|max:50|unique:roles,slug',
@@ -60,6 +64,7 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
+        Gate::authorize('have_access', 'role.show');
         $permission_role = [];
         foreach ($role->permissions as $permission) {
             $permission_role[] = $permission->id;
@@ -76,6 +81,7 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        Gate::authorize('have_access', 'role.edit');
         $permission_role = [];
         foreach ($role->permissions as $permission) {
             $permission_role[] = $permission->id;
@@ -93,7 +99,7 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     {
-        //dd($request->all(), $role);
+        Gate::authorize('have_access', 'role.edit');
         $request->validate([
             'name'          => 'required|max:50|unique:roles,name,' . $role->id,
             'slug'          => 'required|max:50|unique:roles,slug,' . $role->id,
@@ -118,6 +124,7 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        Gate::authorize('have_access', 'role.destroy');
         $role->delete();
         return redirect()->route('role.index')
             ->with('status_success', __('Role successfully removed'));
